crowd-edge-cloud/services/ingest/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
google-cloud-pubsub==2.21.0
pydantic==1.10.14
